Split out StoreBase and RecordId to a separate file (while trying to debug strange crashes with ForeignStore).  Some cleanup to reduce compiler warnings.  Had to clean and recompile - not sure if the issue is resolved now.

+++ apps/openmw/mwclass/creature.cpp
@@ -680,7 +680,7 @@ namespace MWClass
                 ++sound;
             }
             if(!sounds.empty())
-                return sounds[OEngine::Misc::Rng::rollDice(sounds.size())]->mSound;
+                return sounds[OEngine::Misc::Rng::rollDice(static_cast<int>(sounds.size()))]->mSound;
         }
 
         if (type == ESM::SoundGenerator::Land)
